refactor schema primitives to break circular deps
