+++ docs/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Migrated JSON module imports to Node.js 22 import attributes (`with { type: 'json' }`) across engine runtime and test suites to resolve TS2880 and align with the ESM baseline.
 - Normalised workforce trait metadata conflict resolution to avoid mutating read-only maps and aligned identity RNG seeding with branded employee UUIDs for TS 5.4 compatibility.
+- Broke the domain schema circular dependency chain by introducing `schemas/primitives.ts` for shared Zod scalars, moving `InventorySchema` into its own module, and updating every `Uuid`/inventory/harvest import across the engine so inventory parsing no longer dereferences undefined schemas at runtime.
 - Fixed CO₂ injector clamp reporting (Task 0019) and extended tariff bootstrap tests:
   - Corrected `clampedByTarget` so it only flips when the requested delta exceeds deliverable output, and added coverage to verify satisfied requests remain unclamped.
   - Exercised difficulty-specific tariff overrides and cache reuse in `createEngineBootstrapConfig` while syncing SEC/ADR typos with the Node.js 22 baseline.
