+++ packages/engine/src/backend/src/domain/schemas/InventorySchema.ts
@@ -1,13 +1,14 @@
 import { z } from 'zod';
 
-import type { Inventory } from '../types/Inventory.js';
 import { HarvestLotSchema } from './HarvestLotSchema.js';
 
-export const InventorySchema: z.ZodType<Inventory> = z
+export const InventorySchema = z
   .object({
-    lots: z.array(HarvestLotSchema).readonly()
+    lots: z.readonly(z.array(HarvestLotSchema)).default([])
   })
-  .strict();
+  .strip();
+
+export type Inventory = z.infer<typeof InventorySchema>;
 
 export function isInventory(input: unknown): input is Inventory {
   return InventorySchema.safeParse(input).success;
