+++ packages/engine/src/backend/src/domain/types/HarvestLot.ts
@@ -1,4 +1,4 @@
-import type { Uuid } from '../entities.js';
+import type { Uuid } from '../schemas/primitives.js';
 
 export interface HarvestLot {
   readonly id: Uuid;
