+++ packages/engine/src/backend/src/domain/workforce/WorkforceState.ts
@@ -1,4 +1,4 @@
-import type { Uuid } from '../entities.js';
+import type { Uuid } from '../schemas/primitives.js';
 import type { Employee } from './Employee.js';
 import type { EmployeeRole } from './EmployeeRole.js';
 import type { WorkforceKpiSnapshot } from './kpis.js';
