+++ packages/engine/src/backend/src/domain/workforce/intents.ts
@@ -1,4 +1,4 @@
-import type { Uuid } from '../entities.js';
+import type { Uuid } from '../schemas/primitives.js';
 
 export interface WorkforceRaiseAcceptIntent {
   readonly type: 'workforce.raise.accept';
