+++ packages/engine/src/backend/src/domain/workforce/tasks.ts
@@ -1,4 +1,4 @@
-import type { Uuid } from '../entities.js';
+import type { Uuid } from '../schemas/primitives.js';
 import type { EmployeeSkillRequirement } from './EmployeeRole.js';
 
 /**
