+++ packages/engine/src/backend/src/domain/workforce/warnings.ts
@@ -1,4 +1,4 @@
-import type { Uuid } from '../entities.js';
+import type { Uuid } from '../schemas/primitives.js';
 import type { Employee } from './Employee.js';
 import type { WorkforceTaskInstance } from './tasks.js';
