+++ packages/engine/src/backend/src/engine/seedToHarvest.ts
@@ -13,9 +13,9 @@ import type {
   Room,
   SimulationWorld,
   Structure,
-  Uuid,
   Zone
 } from '../domain/world.js';
+import type { Uuid } from '../domain/schemas/primitives.js';
 import { loadStrainBlueprint } from '../domain/blueprints/strainBlueprintLoader.js';
 import type { StrainBlueprint } from '../domain/blueprints/strainBlueprint.js';
